refactor(api): migrate group usecase  to its bounded context
